fix: address reviewer feedback on self-init state machine
